refactor: rename attachments module
