+++ src/views/search/list/message/search-message-list-item-core.tsx
@@ -20,7 +20,7 @@ import { find, includes, isEmpty, reduce } from 'lodash';
 import moment from 'moment';
 import { useTranslation } from 'react-i18next';
 
-import { hasAttachments } from '../../../../attachments';
+import { hasAttachments } from '../../../../attachments/attachments';
 import { getTimeLabel, participantToString } from 'commons/utils';
 import { MailMessage, TextReadValuesType } from 'types/index.d';
 import { useTagExist } from 'ui-actions/tag-actions';
